Extract AnyCodable array extension into separate file
